+++ src/lib.rs
@@ -1,4 +1,7 @@
 use serde::{Deserialize, Serialize};
+use std::io::stdin;
+use std::io::BufRead;
+use std::io::Write;
 
 #[derive(Deserialize, Debug, Serialize, Clone)]
 pub struct Message {
@@ -38,6 +41,33 @@ pub enum Extra {
     InitOk {
         in_reply_to: u64,
     },
+}
+
+pub struct Node {}
+
+impl Node {
+    //TODO generic reader writer;
+    pub fn init<W: Write>(mut writer: W) {
+        let stdin = stdin().lock();
+        let line = stdin
+            .lines()
+            .into_iter()
+            .next()
+            .expect("Init msg always present")
+            .unwrap();
+        let msg: Message = serde_json::from_str(&line).unwrap();
+        serde_json::to_writer(
+            &mut writer,
+            &msg.reply(Body {
+                msg_id: 0,
+                extra: Extra::InitOk {
+                    in_reply_to: msg.body.msg_id,
+                },
+            }),
+        )
+        .unwrap();
+        writer.write_all(b"\n").unwrap();
+    }
 }
 
 #[cfg(test)]
